backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
marshmallow
python-dotenv   
flask-cors
flask-JWT-Extended 